chore: HttpLoadBalancerExtension should expose targetGroup (#278)
Issue Related #209
Here the customer provided use case is not necessarily passing in a target group as a property in the LB constructor, but to expose the target group as an optional field on the Service after it's created. 

So when you add an HTTPLoadBalancerExtension, you can customize the health checks by accessing the targetGroup field on the service.

Here I'm providing an example how to access the targetGroup

```
  const service = new Service(stack, 'my-service', {
  environment,
  serviceDescription,
  autoScaleTaskCount: {
    maxTaskCount: 5,
  },
});

service.targetGroup.configureHealthCheck({
  path: '/',
  port: '80',
});
```

And Here I'm providing yarn compile and yarn test results as shown below:
<img width="1252" alt="image" src="https://user-images.githubusercontent.com/115483524/210608405-59705cf4-8772-460a-8918-ce42ea7a7f65.png">